wrokoutSet and setTemplate controllers adjusted

+++ backend/src/controllers/workoutExercise.controller.ts
@@ -5,7 +5,9 @@ import type { WorkoutExercise } from '../types/workoutExercise.type';
 // GET All WORKOUT EXERCISES
 export const getAllWorkoutExercises = async (req: Request, res: Response) => {
     try {
-        const workoutExercises = await prisma.workoutExercise.findMany({});
+        const workoutExercises = await prisma.workoutExercise.findMany({
+            include: { exercise: true, workout: true, workoutSets: true }
+        });
         res.json(workoutExercises);
     } catch (error) {
         res.status(500).json({ error: "Failed to fetch workout exercises" });
@@ -17,11 +19,13 @@ export const getWorkoutExerciseById = async (req: Request<WorkoutExercise>, res:
     try {
         const { id } = req.params;
         const workoutExercise = await prisma.workoutExercise.findUnique({
-            where: { id: Number(id) }
+            where: { id: Number(id) }, include: { exercise: true, workout: true, workoutSets: true }
         });
+
         if (!workoutExercise) {
             return res.status(404).json({ error: "Workout exercise not found" });
         }
+
         res.json(workoutExercise);
     } catch (error) {
         res.status(500).json({ error: "Failed to fetch workout exercise" });
@@ -29,11 +33,16 @@ export const getWorkoutExerciseById = async (req: Request<WorkoutExercise>, res:
 };
 
 // POST (create a new WORKOUT EXERCISE)
-export const createWorkoutExercise = async (req: Request, res: Response) => {
+export const createWorkoutExercise = async (req: Request<WorkoutExercise>, res: Response) => {
     try {
-        const { workoutId, exerciseId} = req.body;
+        const { workoutId, exerciseId, notes } = req.body;
+
+        if (!workoutId || !exerciseId || !notes) {
+            return res.status(400).json({ error: "Workout ID, Exercise ID, and notes are required" });
+        }
+
         const newWorkoutExercise = await prisma.workoutExercise.create({
-            data: { workoutId, exerciseId }
+            data: { workoutId : workoutId as number, exerciseId : exerciseId as number, notes: notes as string },
         });
         res.status(201).json(newWorkoutExercise);
     } catch (error) {
@@ -45,10 +54,18 @@ export const createWorkoutExercise = async (req: Request, res: Response) => {
 export const updateWorkoutExercise = async (req: Request<WorkoutExercise>, res: Response) => {
     try {
         const { id } = req.params;
-        const { workoutId, exerciseId } = req.body;
+        const { workoutId, exerciseId, notes } = req.body;
+
+        const existingWorkoutExercise = await prisma.workoutExercise.findUnique({
+            where: { id: Number(id) }
+        });
+        if (!existingWorkoutExercise) {
+            return res.status(404).json({ error: "Workout exercise not found" });
+        }
+
         const updatedWorkoutExercise = await prisma.workoutExercise.update({
             where: { id: Number(id) },
-            data: { workoutId: Number(workoutId), exerciseId: Number(exerciseId) }
+            data: { workoutId: Number(workoutId), exerciseId: Number(exerciseId), notes}
         });
         res.json(updatedWorkoutExercise);
     } catch (error) {
@@ -61,6 +78,14 @@ export const updateWorkoutExercise = async (req: Request<WorkoutExercise>, res: 
 export const deleteWorkoutExercise = async (req: Request<WorkoutExercise>, res: Response) => {
     try {
         const { id } = req.params;
+
+        const existingWorkoutExercise = await prisma.workoutExercise.findUnique({
+            where: { id: Number(id) }
+        });
+        if (!existingWorkoutExercise) {
+            return res.status(404).json({ error: "Workout exercise not found" });
+        }
+
         await prisma.workoutExercise.delete({
             where: { id: Number(id) }
         });
